src/qup/tests/consensus_tests.rs
use crate::qup::consensus::QUPConsensus;
use crate::qup::config::QUPConfig;
use crate::qup::crypto::QUPKeyPair;
use crate::qup::state::QUPState;
use std::sync::Arc;

#[test]
fn test_propose_block() {
    // Create a QUPConfig instance
    let config = Arc::new(QUPConfig::default());

    // Create a QUPState instance
    let state = Arc::new(QUPState::new());

    // Create a QUPKeyPair instance
    let key_pair = QUPKeyPair::generate();

    // Create a QUPConsensus instance
    let mut consensus = QUPConsensus::new(config, state, key_pair);

    // Propose a new block with some transactions
    let transactions = vec![]; // Add some transactions here
    let block = consensus.propose_block(transactions).unwrap();

    // Validate the proposed block
    assert!(consensus.validate_block(&block).unwrap());
}

#[test]
fn test_validate_block() {
    // Create a QUPConfig instance
    let config = Arc::new(QUPConfig::default());

    // Create a QUPState instance
    let state = Arc::new(QUPState::new());

    // Create a QUPKeyPair instance
    let key_pair = QUPKeyPair::generate();

    // Create a QUPConsensus instance
    let consensus = QUPConsensus::new(config, state, key_pair);

    // Create a valid block
    let block = /* create a valid block */;

    // Validate the valid block
    assert!(consensus.validate_block(&block).unwrap());

    // Create an invalid block (e.g., with an invalid signature)
    let invalid_block = /* create an invalid block */;

    // Validate the invalid block
    assert!(!consensus.validate_block(&invalid_block).unwrap());
}

// Add more test cases as needed
